Deseasonify: separate seasonal definitions into new module
Since the original `seasons` package is now called just `exts`,
it makes sense to migrate all season-specific definitions out into
a separate module.

This module now contains everything relating to 'new' seasons, i.e.
just branding tooling.
